resolve review comments.

+++ Sources/EkaScribeSDK/API/Models/ScribeSession.swift
@@ -2,13 +2,13 @@ import Foundation
 
 public struct ScribeSession: Sendable {
     public let sessionId: String
-    public let createdAt: Int64
-    public let updatedAt: Int64
+    public let createdAt: Int
+    public let updatedAt: Int
     public let state: String
     public let chunkCount: Int
     public let uploadStage: UploadStage
 
-    public init(sessionId: String, createdAt: Int64, updatedAt: Int64, state: String, chunkCount: Int, uploadStage: UploadStage) {
+    public init(sessionId: String, createdAt: Int, updatedAt: Int, state: String, chunkCount: Int, uploadStage: UploadStage) {
         self.sessionId = sessionId
         self.createdAt = createdAt
         self.updatedAt = updatedAt
