+++ Sources/EkaScribeSDK/API/Models/VoiceActivityData.swift
@@ -3,9 +3,9 @@ import Foundation
 public struct VoiceActivityData: Sendable {
     public let isSpeech: Bool
     public let amplitude: Float
-    public let timestampMs: Int64
+    public let timestampMs: Int
 
-    public init(isSpeech: Bool, amplitude: Float, timestampMs: Int64) {
+    public init(isSpeech: Bool, amplitude: Float, timestampMs: Int) {
         self.isSpeech = isSpeech
         self.amplitude = amplitude
         self.timestampMs = timestampMs
