+++ Sources/EkaScribeSDK/Common/TimeProvider.swift
@@ -1,11 +1,11 @@
 import Foundation
 
 protocol TimeProvider {
-    func nowMillis() -> Int64
+    func nowMillis() -> Int
 }
 
 struct DefaultTimeProvider: TimeProvider {
-    func nowMillis() -> Int64 {
-        Int64(Date().timeIntervalSince1970 * 1000)
+    func nowMillis() -> Int {
+        Int(Date().timeIntervalSince1970 * 1000)
     }
 }
