+++ Sources/EkaScribeSDK/Recorder/AudioFrame.swift
@@ -2,7 +2,7 @@ import Foundation
 
 struct AudioFrame: Sendable {
     let pcm: [Int16]
-    let timestampMs: Int64
+    let timestampMs: Int
     let sampleRate: Int
-    let frameIndex: Int64
+    let frameIndex: Int
 }
